quizforge/requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4
reportlab==4.0.7
fpdf2==2.7.6
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
nltk==3.8.1
